Fixes for security holes in web services - ISSUE 419.
git-svn-id: http://ala-fieldcapture.googlecode.com/svn/trunk/FieldCapture@1155 63d11371-6d71-872f-44d1-802e8bf8a75e

+++ grails-app/controllers/au/org/ala/fieldcapture/UserController.groovy
@@ -82,9 +82,14 @@ class UserController {
         String userId = params.userId
         String role = params.role
         String projectId = params.projectId
+        def adminUser = authService.userDetails()
 
-        if (projectId && role && userId) {
-            render userService.removeUserWithRole(projectId, userId, role) as JSON
+        if (adminUser && projectId && role && userId) {
+            if (projectService.isUserAdminForProject(adminUser.userId, projectId)) {
+                render userService.removeUserWithRole(projectId, userId, role) as JSON
+            } else {
+                render status:403, text: 'Permission denied'
+            }
         } else {
             render status:400, text: 'Required params not provided: userId, projectId, role'
         }
